# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Payment Gateway Merchant API
  version: 1.0.0
  description: |
    Server-to-server API for merchants. Create a payment, confirm it with a payment method, and follow
    `next_action` until the payment reaches a final status. Alternatively, create a checkout session and
    redirect the customer to its hosted page.

    - **Authentication**: `Authorization: Bearer sk_...`. The key is a server-side secret; never ship it to a
      browser or an app.
    - **Idempotency**: every `POST` requires an `Idempotency-Key` (1-255 visible ASCII characters). A retry with
      the same key and body returns the original response with `Idempotent-Replayed: true`. The same key with a
      different body is rejected with `422 idempotency_key_reuse`. Keys are kept for 7 days.
    - **Format**: JSON in snake_case. Amounts are integers in minor units (paise for INR). Timestamps are RFC 3339
      in UTC. Fields without a value are omitted rather than sent as `null`. Clients must ignore unknown fields.
    - **Errors**: RFC 9457 problem details (`application/problem+json`) with a stable machine-readable `code`.
    - **Rate limits**: separate per-merchant budgets for reads (`GET`) and writes. Over budget you get
      `429 rate_limited` with `Retry-After`; the request was not processed and did not consume its
      `Idempotency-Key`, so retry it unchanged.
...
      on webhooks (see `webhooks`) or `GET /v1/payments/{payment_id}` for the final status.
...

...
servers:
...
security:
  - apiKey: []
...
paths:
...
  /v1/mandates:
    post:
      tags: [Mandates]
      operationId: createMandate
      summary: Register a mandate
      description: |
        Registers the mandate at the first linked PSP that supports the instrument and `max_amount`. The response
        is usually `pending_authorization` with a `next_action` (`redirect`) where the customer approves it within
        24 hours. UPI and card registrations charge ₹1 through the payment in `registration_payment_id`. The
        outcome arrives as `mandate.activated` or `mandate.failed`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateMandateRequest'}
            example:
              instrument: upi_autopay
              max_amount: 500000
              currency: INR
              frequency: monthly
              description: Gym membership
              customer: {reference: cust_42, name: Asha Rao, email: asha@example.com, phone: '+919999999999'}
              return_url: https://merchant.example/subscriptions/42
      responses:
        '201':
          description: Mandate created (or the original response replayed).
          headers:
            Idempotent-Replayed: {$ref: '#/components/headers/IdempotentReplayed'}
            X-Request-Id: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Mandate'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
        default: {$ref: '#/components/responses/UnexpectedError'}
  /v1/mandates/{mandate_id}:
    get:
      tags: [Mandates]
      operationId: getMandate
      summary: Retrieve a mandate
      parameters:
        - $ref: '#/components/parameters/MandateId'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: The mandate.
          headers:
            X-Request-Id: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Mandate'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        default: {$ref: '#/components/responses/UnexpectedError'}
  /v1/mandates/{mandate_id}/revoke:
    post:
      tags: [Mandates]
      operationId: revokeMandate
      summary: Revoke a mandate
      description: |
        Revokes the mandate at the PSP, then here, and cancels its debits that have not started executing. A revoked
        mandate replays as success. `409 mandate_invalid_state` once it has expired or failed; `503` when the PSP
        could not confirm the revocation (retry with the same Idempotency-Key).
      parameters:
        - $ref: '#/components/parameters/MandateId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: The revoked mandate.
          headers:
            Idempotent-Replayed: {$ref: '#/components/headers/IdempotentReplayed'}
            X-Request-Id: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Mandate'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        '503': {$ref: '#/components/responses/ServiceUnavailable'}
        default: {$ref: '#/components/responses/UnexpectedError'}
  /v1/mandates/{mandate_id}/debits:
    post:
      tags: [Mandates]
      operationId: createMandateDebit
      summary: Debit an active mandate
      description: |
        Schedules a debit. For UPI AutoPay and card mandates the PSP notifies the customer, and the debit runs at
        least 24 hours after delivery and no earlier than `due_at`. A failed debit is retried up to 3 times, a day
        apart. The money moves on the debit's payment (`payment_id`): follow its `payment.*` events. A mandate has
        one debit in progress at a time. UPI and card debits above ₹15,000 (or your account's limit) need the
        customer's authentication, so collect them as payments instead.
      parameters:
        - $ref: '#/components/parameters/MandateId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/CreateMandateDebitRequest'}
            example:
              amount: 99900
              merchant_debit_id: invoice_2026_10
              description: October membership
      responses:
        '201':
          description: Debit scheduled (or the original response replayed).
          headers:
            Idempotent-Replayed: {$ref: '#/components/headers/IdempotentReplayed'}
            X-Request-Id: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MandateDebit'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        default: {$ref: '#/components/responses/UnexpectedError'}
    get:
      tags: [Mandates]
      operationId: listMandateDebits
      summary: List the debits of a mandate
      parameters:
        - $ref: '#/components/parameters/MandateId'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: Debits, oldest first.
          headers:
            X-Request-Id: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MandateDebitList'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        default: {$ref: '#/components/responses/UnexpectedError'}
  /v1/mandates/{mandate_id}/debits/{debit_id}:
    get:
      tags: [Mandates]
      operationId: getMandateDebit
      summary: Retrieve a debit
      parameters:
        - $ref: '#/components/parameters/MandateId'
        - $ref: '#/components/parameters/DebitId'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: The debit.
          headers:
            X-Request-Id: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MandateDebit'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        default: {$ref: '#/components/responses/UnexpectedError'}
  /v1/mandates/{mandate_id}/debits/{debit_id}/cancel:
    post:
      tags: [Mandates]
      operationId: cancelMandateDebit
      summary: Cancel a debit
      description: |
        Cancels a debit that has not started executing (`scheduled`, `notifying` or `ready`), and its payment. A
        cancelled debit replays as success; any other status is `409 mandate_invalid_state`.
      parameters:
        - $ref: '#/components/parameters/MandateId'
        - $ref: '#/components/parameters/DebitId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestId'
      responses:
        '200':
          description: The cancelled debit.
          headers:
            Idempotent-Replayed: {$ref: '#/components/headers/IdempotentReplayed'}
            X-Request-Id: {$ref: '#/components/headers/RequestId'}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/MandateDebit'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '404': {$ref: '#/components/responses/NotFound'}
        '409': {$ref: '#/components/responses/Conflict'}
        '422': {$ref: '#/components/responses/Unprocessable'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
        default: {$ref: '#/components/responses/UnexpectedError'}
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      description: Secret API key (`sk_...`), issued per merchant.
  parameters:
...
    MandateId:
      name: mandate_id
      in: path
      required: true
      schema: {type: string, pattern: '^mdt_'}
    DebitId:
      name: debit_id
      in: path
      required: true
      schema: {type: string, pattern: '^mdd_'}
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique per logical operation; reuse it only to retry the same request.
      schema: {type: string, minLength: 1, maxLength: 255, pattern: '^[\x21-\x7E]+$'}
    RequestId:
      name: X-Request-Id
      in: header
      required: false
      description: Optional correlation id (1-64 of `A-Za-z0-9._-`); generated when absent and echoed back.
      schema: {type: string, pattern: '^[A-Za-z0-9._-]{1,64}$'}
  headers:
    RequestId:
      required: true
      description: Correlation id of this request; quote it when contacting support.
      schema: {type: string}
    IdempotentReplayed:
      required: true
      description: '`true` when this is the stored response of an earlier request with the same key.'
      schema: {type: string, enum: ['true', 'false']}
    RetryAfter:
      description: Seconds to wait before retrying.
      schema: {type: integer, minimum: 1}
    RetryAfterRequired:
      required: true
      description: Seconds to wait before retrying.
      schema: {type: integer, minimum: 1}
  responses:
    BadRequest:
      description: '`validation_error` (see `errors`), `malformed_request` or `idempotency_key_required`.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthorized:
      description: '`authentication_required` or `invalid_api_key`.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: '`resource_not_found`: unknown id, or it belongs to another merchant.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Conflict:
      description: |
        `payment_invalid_state` (re-read the resource), `refund_already_exists`, `mandate_invalid_state`,
        `mandate_debit_already_exists`, or `idempotency_request_in_progress` (retry with the same key after
        `Retry-After`).
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unprocessable:
      description: |
        `idempotency_key_reuse`, `amount_exceeds_refundable`, `amount_exceeds_mandate_limit`,
        `capture_amount_mismatch`, `unsupported_payment_method` or `unsupported_currency`. Do not retry unchanged.
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    TooManyRequests:
      description: '`rate_limited`. Not processed and the Idempotency-Key was not consumed: retry after `Retry-After`.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
        Retry-After: {$ref: '#/components/headers/RetryAfterRequired'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    ServiceUnavailable:
      description: '`no_provider_available`: retry with the same Idempotency-Key and backoff.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    UnexpectedError:
      description: '`internal_error`: retry with the same Idempotency-Key and backoff.'
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
    Amount:
      type: integer
      format: int64
      description: Minor units of `currency` (paise for INR).
    Timestamp:
      type: string
      format: date-time
    Currency:
      type: string
      pattern: '^[A-Z]{3}$'
      examples: [INR]
...
    FailureCategory:
      type: string
      enum: [customer, issuer, provider, provider_unavailable, validation, risk, timeout, not_submitted]
...
    MandateStatus:
      type: string
      enum: [created, pending_authorization, active, paused, revoked, expired, failed]
      description: |
        Only `active` mandates can be debited. `paused` and `active` alternate as the customer pauses and resumes in
        their UPI or bank app. Final: `revoked`, `expired` (at `end_at`) and `failed` (never authorized).
    MandateDebitStatus:
      type: string
      enum: [scheduled, notifying, ready, executing, succeeded, failed, cancelled]
      description: |
        UPI and card debits go `scheduled` → `notifying` → `ready` → `executing`; eNACH debits start `ready`.
        A failed attempt with retries left starts the next cycle. Final: `succeeded`, `failed` and `cancelled`.
    MandateInstrument:
      type: string
      enum: [upi_autopay, card, enach]
    MandateFrequency:
      type: string
      enum: [daily, weekly, fortnightly, monthly, bimonthly, quarterly, half_yearly, yearly, as_presented]
      description: Passed to the PSP, which enforces it.
...
    NextAction:
      type: object
      description: Present only while the payment is `requires_action`.
      required: [type]
      properties:
        type: {type: string, enum: [redirect, upi_intent, display_qr, await_approval]}
        url: {type: string, description: '`redirect`: send the customer to this URL (a PSP-hosted page).'}
        upi_uri: {type: string, description: '`upi_intent`: open this `upi://` URI on the customer''s phone.'}
        qr_payload: {type: string, description: '`display_qr`: render this as a QR code.'}
        expires_at: {$ref: '#/components/schemas/Timestamp'}
...
    PaymentError:
      type: object
      required: [code]
      properties:
        code: {type: string}
        category: {$ref: '#/components/schemas/FailureCategory'}
        message: {type: string}
...
    Mandate:
      type: object
      required: [id, object, status, instrument, provider, max_amount, currency, frequency, start_at, end_at,
                 customer, created_at, updated_at, version]
      properties:
        id: {type: string, pattern: '^mdt_'}
        object: {const: mandate}
        status: {$ref: '#/components/schemas/MandateStatus'}
        instrument: {$ref: '#/components/schemas/MandateInstrument'}
        provider: {type: string}
        max_amount:
          $ref: '#/components/schemas/Amount'
          description: The largest single debit.
        currency: {$ref: '#/components/schemas/Currency'}
        frequency: {$ref: '#/components/schemas/MandateFrequency'}
        start_at: {$ref: '#/components/schemas/Timestamp'}
        end_at: {$ref: '#/components/schemas/Timestamp'}
        description: {type: string}
        customer: {$ref: '#/components/schemas/MandateCustomer'}
        metadata:
          type: object
          additionalProperties: {type: string}
        next_action:
          $ref: '#/components/schemas/NextAction'
          description: Present only while `pending_authorization`; send the customer there to approve the mandate.
        registration_payment_id:
          type: string
          pattern: '^pay_'
          description: The ₹1 authorization charge of UPI and card mandates.
        last_error: {$ref: '#/components/schemas/PaymentError'}
        activated_at: {$ref: '#/components/schemas/Timestamp'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
        version: {type: integer, format: int64}
    MandateCustomer:
      type: object
      required: [email, phone]
      properties:
        reference: {type: string}
        name: {type: string}
        email: {type: string}
        phone: {type: string}
    MandateDebit:
      type: object
      description: One debit of a mandate. Its payment carries the money and the outcome.
      required: [id, object, mandate_id, payment_id, merchant_debit_id, amount, currency, status, due_at, not_before,
                 cycle, created_at, updated_at, version]
      properties:
        id: {type: string, pattern: '^mdd_'}
        object: {const: mandate_debit}
        mandate_id: {type: string, pattern: '^mdt_'}
        payment_id: {type: string, pattern: '^pay_'}
        merchant_debit_id: {type: string}
        amount: {$ref: '#/components/schemas/Amount'}
        currency: {$ref: '#/components/schemas/Currency'}
        status: {$ref: '#/components/schemas/MandateDebitStatus'}
        description: {type: string}
        due_at: {$ref: '#/components/schemas/Timestamp'}
        not_before:
          $ref: '#/components/schemas/Timestamp'
          description: Earliest execution of the current cycle.
        cycle: {type: integer, minimum: 1, description: '1 for the first attempt, then one more per retry.'}
        notified_at:
          $ref: '#/components/schemas/Timestamp'
          description: When the PSP delivered the current cycle's pre-debit notification.
        last_error: {$ref: '#/components/schemas/PaymentError'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        updated_at: {$ref: '#/components/schemas/Timestamp'}
        version: {type: integer, format: int64}
    MandateDebitList:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: {$ref: '#/components/schemas/MandateDebit'}
...
    Problem:
      type: object
      required: [title, status, code]
      properties:
        type: {type: string}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        code:
          type: string
          enum: [validation_error, malformed_request, idempotency_key_required, authentication_required,
                 invalid_api_key, invalid_signature, resource_not_found, method_not_allowed, unsupported_media_type,
                 payment_invalid_state, refund_already_exists, mandate_invalid_state, mandate_debit_already_exists,
                 idempotency_request_in_progress, idempotency_key_reuse, amount_exceeds_refundable,
                 amount_exceeds_mandate_limit, capture_amount_mismatch, unsupported_payment_method,
                 unsupported_currency, rate_limited, no_provider_available, internal_error]
        request_id: {type: string}
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: {type: string}
              message: {type: string}
...
    CreateMandateRequest:
      type: object
      required: [instrument, max_amount, currency, frequency, customer]
      properties:
        instrument: {$ref: '#/components/schemas/MandateInstrument'}
        max_amount:
          type: integer
          format: int64
          minimum: 100
          maximum: 1000000000
          description: The largest single debit; the PSP's limit for the instrument applies.
        currency: {$ref: '#/components/schemas/Currency'}
        frequency: {$ref: '#/components/schemas/MandateFrequency'}
        start_at:
          $ref: '#/components/schemas/Timestamp'
          description: Defaults to now.
        end_at:
          $ref: '#/components/schemas/Timestamp'
          description: Defaults to 10 years after `start_at`.
        description: {type: string, maxLength: 255}
        customer:
          type: object
          required: [email, phone]
          properties:
            reference: {type: string, maxLength: 64}
            name: {type: string, maxLength: 100}
            email: {type: string, format: email, maxLength: 254}
            phone: {type: string, pattern: '^\+?[0-9]{8,15}$'}
        metadata:
          type: object
          maxProperties: 20
          propertyNames: {minLength: 1, maxLength: 40}
          additionalProperties: {type: string, maxLength: 500}
        return_url:
          type: string
          maxLength: 2048
          pattern: '^https?://\S+$'
          description: Where the PSP's authorization page sends the customer back.
    CreateMandateDebitRequest:
      type: object
      required: [amount, merchant_debit_id]
      properties:
        amount: {type: integer, format: int64, minimum: 100, maximum: 1000000000}
        merchant_debit_id:
          type: string
          maxLength: 64
          pattern: '^[A-Za-z0-9_\-:.]+$'
          description: Your reference, unique per mandate; also the payment's `merchant_order_id`.
        due_at:
          $ref: '#/components/schemas/Timestamp'
          description: 'Earliest debit. Default: 25 hours from now for UPI and card (after the notice), now for eNACH.'
        description: {type: string, maxLength: 255}
...